Allow to disable the default keybindings
Usually I do not like it when plugins just setup their own keybindings,
in particular when it overwrites mappings either from vim's default
bindings or from the user's config.

In this case an exception can be made, because F1 is ubiquitously used
for help and the default keybinding just replicates :h, which is not
terribly useful, so overwriting it should not be terribly controversial.

Just in case however, this patch still allows to disable the default
keybindings.