Fix off-by-one if word at end of line
If the word goes until the end of the line, then the searchpos() forward
search does not find a non-matching character and returns [0,0]. The
"-2" then does not work for this special case.

In general "-2" is correct, as we need to subtract one for converting
from line-space to array-index-space and another one because we have to
go one character left from the first non-matching one.

But in this edge case if there is no non-matching character, we end up
out of bounds. Luckily we can fix this by simply lifting "-2" to "-1",
because "-1" also refers to the last character.