Walk docs in a hierarchical way
With this heuristic we are most likely to get the correct docset as the
first docset, which is important if the query matches multiple docsets
at once, e.g. std::string matches both Rust and C++.

The only design problem with this is that it does not keep the order in
additional_docs(), since it gets lost in the lambda which only checks if
it matches. Instead the order is given by the way that we iterate over
s:docs. In the future it would be nice to instead keep the order from
additional_docs(), but for now this is fine.