Add fzf support
This allows to fuzzy-match against all sites in a docset at once.

The fzf wrapper is a bit cursed for multiple reasons.

First of all the, "--with-nth" argument implementation in fzf for
selecting an entry when mapping items to a custom string [0] is
absolutely retarded, because it keeps the separators.
For example consider:
echo 'foo|bar' | fzf --delimiter '\|' --with-nth 1

This displays "foo|" instead of the expected "foo".

Luckily we can work around this weird-ass bizarre inclusion of the
separator by choosing a non-visible separator such as "\t", so that the
user does not see this dumpster fire [1]. And luckily fzf by default
already strips whitespace around all field index expressions.

The next shitshow is that it is not possible at all to field-delimit the
returned value when the user presses Enter, so we always get back the
entire string with all fields included.

For now we render the preview with pandoc. It would be wise to strip the
unnecessary header for some pages so that the user does not need to
scroll down.
The next steps are to add keybindings to open the documentation in
various ways.

[0] https://github.com/junegunn/fzf/issues/3316
[1] https://github.com/junegunn/fzf/issues/3529#issuecomment-1844991143