Open documentation on Enter
Once again [0], the implementation details in fzf are completely bogus.

For some reason the maintainer believes so much in the Unix philosophy
that field index expressions are not allowed in the final returned
string [1], which means that the final string we get in the sink is some
bullshit like "std::string|/path/to/doc.html".  Apparently Unix nerds
will automatically jump to awk or cut to handle this stuff.

To troll the fzf maintainer, let's use VimScript instead of proper Unix
tools.

[0] 00022ad8d5c42830e3e2064e2533ca2a3abee8e2
[1] https://github.com/junegunn/fzf/issues/1323