Add neovim checkhealth support
WTF Lua are you actually fucking retarded?
To count the numbers of a table, the "#" operator actually only works if
the table is a contiguous array. Otherwise it just returns random
garbage [0].
The stackoverflow answer unironically suggests iterating over the table
to compute the count, RIP runtime complexity. Why are people suggesting
using Lua over Vimscript again?
Who would have thought that this whole map/array dichotomy would cause
problems on the long run...

At least there is a helper function in the vim.tbl namespace, but this
should be a Lua built-in, there is just no excuse.

[0] https://stackoverflow.com/a/2705804