Add a function to show docs with elinks
The special treat about this popup frontend is that it features a much
more sophisticated algorithm for skipping to anchors, than the pandoc
frontend one.

While the pandoc code essentially just skips ahead until it finds a
substring match with the anchor, the new CSS visibility support in
elinks [0] allows us to render everything before the anchor invisible
with pure CSS. A simple strip()-style post-processing step than allows
us to skip right ahead to the anchor.

Note that resolving the anchor to an id can sometimes be tricky,
because some websites like to add some special long URLs like
"//apple_ref/Function/setUnicode", but usually we can get away with
stripping out the last element of this path, e.g. in this case an
element with id "setUnicode" does exist.

For now this new functionality is not used anywhere, as we need to wait
a little bit until it makes it into an elinks release.
Regardless, there are still some things to polish before this can become
default.

[0] https://github.com/rkd77/elinks/issues/341